add _isDestroyed flag that prevents resubscribing to status stream (#158)
Currently in the client SDK, destroying a status stream with `destroy()`
triggers the clause for detecting unexpected stream terminations and
recreates the status stream. This causes an error some time later when
the hanging stream times out.


![image](https://github.com/user-attachments/assets/16ec174f-5d2e-425d-bd07-ea48f67c1c44)

This change simply adds a flag, `_isDestroyed` to prevent the stream
recreation in the case we actually did intend to close it!